MP Improvements (2/4): * Tilemaps now default to general and are less "magic number-y" * Blocks default to being minable with a pickaxe, and are too less magic numbery * Fixed crafting in two ways - the itemstack held by the mouse is not size 1x1 anymore and crafting it not blocked entirely after multiple players logon to the server. * Added a rudimentary block to prevent outdated clients from connecting to a server. This will improve in the next patch. * Metadata autogenerates for sizes upto 100x100 (up from 3x3) * Fixed a really wierd Block duplication bug with blocks of size greater than 1x1, where each part of the block could be mined. * Affix generation should now be nullpointer and bounds safe. Probably. * Reduced the size of save files by ~25% + Documentation * Removed some old classes

+++ TerraeRasa/Server/src/savable/SavableBlock.java
@@ -9,8 +9,7 @@ public class SavableBlock
 {
 	private static final long serialVersionUID = 1L;
 	public ItemStack[] mainInventory;
-	public int id;
-	public int metaData;
-	public int bitMap;
-	
+	public short id;
+	public byte metaData;
+	public byte bitMap;
 }
